All tasks are now created immediately, rather than in `project.afterEvaluate`, for easier management by other scripts.
