+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/SelfTest.java
@@ -122,7 +122,6 @@ public class SelfTest {
 		SpotlessPlugin plugin = project.getPlugins().apply(SpotlessPlugin.class);
 		// setup the plugin
 		test.accept(plugin.getExtension());
-		plugin.createTasks(project);
 		// return the configured plugin
 		return project;
 	}
